car_prices_tool_website/requirements.txt
asgiref==3.3.1
attrs==20.3.0
Click==7.0
Django==3.1.7
django-debug-toolbar==3.2
djangorestframework==3.12.2
Faker==6.4.1
iniconfig==1.1.1
mixer==7.1.2
packaging==20.9
pendulum==2.1.2
pip-tools==5.5.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-decouple==3.4
whitenoise==5.2.0
